unused-requirements.txt
fsspec
jupyterlab
python-json-logger
pyasn1
asttokens
cachetools
nbclient
smart-open
gast
tensorboard-data-server
wcwidth
referencing
urllib3
joblib
fuzzywuzzy
contourpy
jedi
pywinpty
defusedxml
six
widgetsnbextension
absl-py
rsa
cycler
requests
prometheus-client
overrides
torch
rfc3986-validator
opt-einsum
zipp
sympy
pyasn1-modules
idna
bleach
distlib
errant
nbconvert
tensorboard
optree
uri-template
notebook
pandocfilters
webencodings
pipenv
flatbuffers
python-dateutil
debugpy
ipython
exceptiongroup
pyenchant
pyzmq
rpds-py
typer
psutil
beautifulsoup4
namex
pycparser
cryptography
transformers
opencv-python
tornado
jupyter_client
wasabi
preshed
parso
itsdangerous
keras
spacy-loggers
asgiref
ml-dtypes
scipy
spacy
importlib_metadata
prompt-toolkit
matplotlib
jsonpointer
tensorflow-io-gcs-filesystem
cloudpathlib
jupyter-events
billiard
jsonschema-specifications
amqp
fastjsonschema
terminado
safetensors
virtualenv
scikit-learn
async-lru
jupyter
vine
mistune
json5
jupyter_server
pyparsing
termcolor
thinc
executing
charset-normalizer
srsly
tensorflow-estimator
fqdn
ipywidgets
soupsieve
importlib-resources
tomli
typing_extensions
wrapt
huggingface-hub
confection
nbformat
argon2-cffi
markdown-it-py
google-auth-oauthlib
packaging
click-didyoumean
jupyterlab-widgets
tokenizers
arrow
libclang
fonttools
threadpoolctl
annotated-types
blinker
webcolors
anyio
certifi
pydantic
jupyter-console
tensorflow-intel
seaborn
pycryptodome
notebook_shim
cffi
ipykernel
jupyter-lsp
spacy-legacy
kiwisolver
murmurhash
stack-data
kombu
pywin32
protobuf
sniffio
regex
platformdirs
shellescape
tzdata
pandas
google-pasta
rapidfuzz
numpy
mpmath
requests-oauthlib
rich
grpcio
django-admin-honeypot
sqlparse
jupyterlab_pygments
jupyterlab_server
tinycss2
filelock
h5py
blis
mdurl
traitlets
jsonschema
networkx
isoduration
nest-asyncio
jupyter_core
sentencepiece
pydantic_core
click-repl
colorama
astunparse
async-timeout
catalogue
matplotlib-inline
weasel
qtconsole
oauthlib
langcodes
jupyter_server_terminals
pure-eval
wordcloud
google-auth
click-plugins
pyspellchecker
tqdm
cymem
rfc3339-validator
